pyproject.toml
[project]
name = "qa-api"
version = "1.0.0"
description = "Question-Answer API with vector search and LLM generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.1",
    "uvicorn[standard]>=0.38.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "asyncpg>=0.30.0",
    "psycopg>=3.2.12",
    "pinecone>=7.3.0",
    "pinecone-text>=0.11.0",
    "sentence-transformers>=5.1.2",
    "google-generativeai>=0.8.5",
    "google-genai>=1.50.0",
    "langchain>=1.0.5",
    "langchain-core>=1.0.4",
    "langchain-community>=0.4.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.0.0",
    "anyio>=4.0.0",
    "numpy>=1.24.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "database", "src", "settings"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

